refactor(types): move config types to a separate file and adjust imports
- Extract WorkspaceConfig and WorkspaceConfigItem types to src/types/config.ts
- Update imports in cmds and services to use the new types location
- Remove type definitions from config-manager.ts and import them instead
- Adjust import ordering and grouping in workspace-manager.ts
- Modify VSCode settings to always organize imports on save for consistency
