+++ .vscode/settings.json
@@ -5,7 +5,7 @@
         "editor.defaultFormatter": "denoland.vscode-deno",
         "editor.codeActionsOnSave": {
             "source.fixAll": "explicit",
-            "source.organizeImports": "never"
+            "source.organizeImports": "always"
         },
         "editor.tabSize": 4
     },
