+++ src/cmds/enable.ts
@@ -3,8 +3,9 @@ import { blue, green, red, yellow } from "@std/fmt/colors";
 import { Result } from "typescript-result";
 import { ErrorWithCause } from "../libs/errors.ts";
 import { isDir } from "../libs/file.ts";
-import { ConfigManager, type WorkspaceConfig, type WorkspaceConfigItem } from "../services/config-manager.ts";
+import { ConfigManager } from "../services/config-manager.ts";
 import { InteractivePrompt } from "../services/interactive-prompt.ts";
+import type { WorkspaceConfig, WorkspaceConfigItem } from "../types/config.ts";
 import { syncCommand } from "./sync.ts";
 
 export type EnableCommandOption = {
